Helper functions to ensure MIS works in 32-bit and 64-bit
The MIS containers pointers, which will change size when we switch
from 32-bit to 64-bit. This would make the MIS unreadable in 64-bit
mode. To work around this, we test for the architecture and either
define the struct with real pointers, or define it with fixed-size
integers. We then provide helper functions that will always return
valid pointers, both in 32-bit and 64-bit modes.