Suppress a bunch of pointer-to-int-cast warnings
Most of these pointers are being cast to uint64_t for arithmetic,
which is somewhat necessary since ELF headers use 64-bit values and
64-bit addresses, but this produces warnings because we're in 32-bit
mode and the pointer size does not match. If we cast it to a uintptr_t
first, the result is exactly the same, but it suppresses the warning
by first casting to a type of (guaranteed) the same size.