Isolate printf from panic.c into libk
Currently doesn't produce a valid return code. We aren't using it at
the moment, but the term_print_* functions will need to be adjusted to
ensure it's doable. Equally, all of the terminal-specific functions
should be refactored out into puts and putchar functions, which will
make printf independent of how characters are written to the
screen (and, importantly, not libk-specific!)