Create flags struct and implement flags in int printing functions
flags struct makes it a lot simpler to pass the flags to each of the
printing functions. The most basic flags (which don't require other
features, e.g. width) have been implemented: int printing forces a +
sign or a space when '+' or ' ' are passed (when both are passed, the
+ takes precedence and the space is ignored); when '#' is passed, the
hex formatter adds 0x or 0X dependent on case.