Basic working strtoumax and strtoimax (base 10 only)
Goes via a common strtoany function that should be easy to extend to
the other cases. The trick here is specifying whether or not the max
value is that of a signed or unsigned type; this allows the correct
extreme value to be returned if an overflow should occur by exploiting
two's complement, i.e. INTMAX_MAX = abs(INTMAX_MIN)-1

Created some simple tests around the extrema, and the implementation
appears to behave correctly.