Relocate kernel to very high memory, fixing issues that come up
We get A LOT of symbol relocation errors if we don't impose
-mcmodel=large when compiling. This applies in libk as well as the
kernel itself. Additionally, setting up the stack is no longer
possible from C for the same reason; instead, the GCC inline ASM entry
function needs to be relocated to a pure NASM function, and the space
and symbols for the stack need to be created by the linker script
instead. All of this together, however, allows the kernel to be
relocated to 0xfffffe8000200000, i.e. 2MB into the third-to-last entry
of the PML4T.