Refactor the paging code into its own file
Doing this has the weird side effect of causing the compiler to start
using SIMD instructions in elf64_map_program_image. This doesn't work
in this early environment because we haven't enabled the necessary CR0
and CR4 flags for SSE, and it wouldn't surprise me if there was
additionally an alignment error, since it's raising #GP and not #UD.
Nonetheless, the solution is to disable SSE and MMX at the
compiler-level; I believe this is good practice anyway (Linux
generally doesn't permit SSE/MMX in the kernel) except in rare
circumstances because it is necessary to use the (expensive) FXSAVE
and FXRSTOR instructions to preserve the SSE, FPU and MMX states on
return to userspace.