Remove volatile qualifiers
The main reason these were in place was because of zero_page_table
being optimised more heavily in the newer GCC, and creating a general
protection fault in the process. Using volatile preventing
optimisation of these memory accesses, but we now know that the issue
was caused by SSE/MMX. Now these are disabled, we can allow the
compiler to take proper control over these variables.