Fix bump allocator init address
The __loader_end symbol in the linker script essentially becomes a
variable defined at the end of the ELF. The contents of this variable
is meaningless and essentially undefined; the only thing that matters
is the address of this variable. This fixes the highest address
calculation code to use the *address* of __loader_end, rather than the
value, which ensures the bump allocator starts at the correct
address. The previous code caused undefined behaviour that would (if
the bump loader was used excessively) eventually overwrite the module
loader stack and the statically-allocated page tables.