Macros to generate handler stubs, then redirect to a global handler
Global handler should eventually be a switch statement that redirects
to the individual logic, but for now just panics regardless of
interrupt. Macro generates a stub for each handler that just pushes
the interrupt number, then jumps to the appropriate course of
action (either the interrupt pushes an error code, or we need to
normalise the stack frme by pushing one ourselves).