Compile with -mno-red-zone to prevent potential pitfalls
See https://wiki.osdev.org/Libgcc_without_red_zone for more
information. The sysv ABI permits a 128-byte "red zone" that the
compiler can use freely to e.g. avoid creating call frames. This is
fine in userspace code, but if the kernel uses the red zone when an
interrupt triggers, it will clobber the red zone and create
problems. The solution is just to disable the red zone in the kernel.

libgcc by default will be compiled to use the red zone. To get around
this, you need a patched compiler with a version of libgcc that
doesn't use the red zone. This can be set up using multilib. See
https://wiki.osdev.org/Libgcc_without_red_zone for instructions.
See also https://wiki.osdev.org/Talk:Libgcc_without_red_zone for a
discussion of why we need a patched compiler and can't just use
CFLAGS_FOR_TARGET for this -- essentially, patching/multilib ensures
crtbegin and crtend are both built with -mno-red-zone as well, but
this is not necessarily true when doing it via CFLAGS_FOR_TARGET.