Early, basic, single-page PMM
Can currently allocate only one page at a time. Uses a uint64_t to
scan for free pages 64 pages at a time. When a 64-page block is found
with a free page in it, uses ctz to efficiently find the offest to
that page. This method only works when the bitmap stores a single bit
per page, so the implementation is gated; this may not change in the
future, but it's worth being aware of and locking it behind a
preprocessor gate.

Once a free page is found, we mark it as used in the bitmap and return
the physical address of the page. Pages are freed by putting in the
address of the page (and the number of contiguous pages, for
compatibility with the future) and marking those pages as free in the
bitmap.

There is also preliminary code for a basic optimisation. We keep track
of the lowest page marked as free and start from there when searching
for free pages. This is especially useful since the first ~2MB are
immediately, permanently reserved by the PMM, so it would be a waste
of time to scan these pages. Plus, if the page is actually free, we
can also skip scanning for free pages and just immediately allocate
it.