Fix interrupt handlers that don't have error codes
When an interrupt didn't include its own error code, we push an error
code ourselves to normalise the stack frame. However, I was pushing
the error code _after_ the interrupt number instead of before it,
which was causing every no-error-code interrupt to be identified as a
divide error instead of the true interrupt type. This fixes the macro
to push things in the correct order.