Create functions for walking the recursive entry
get_page_table_pointer is used to get a pointer to a specific element
of the page table at a given layer. You *can* use this function to
walk the page table, but it's a little unintuitive. It is intended to
form the backend of the four helper functions provided. These allow
you to retrieve addresses for the start of the PML4T/a particular
PDPT/PD/PT and then treat this as an array of 512 PageTableEntry
elements. These are much more intuitive to work with.