Datatype adjustments
Page table index will not exceed 511, so only 9 bits are needed to
store it. The C standard guarantees that an unsigned short will be at
least 16 bits long, which is the next biggest int size after char. We
therefore use this as a generic type for the index, with the hope that
the compiler can choose the optimal size for the architecture based on
this guaranteed minimum.

This type is also set as a type alias, so it can easily be adjusted
later on.