Basic algorithm for mapping a contiguous free virtual address range
Most of the algorithm is now in place, but the final stage where the
mapping is actually created still needs to be written.

is_virt_addr_used currently just returns whether a page table is
present, also taking large pages into account. This can be modified
later for whatever method the kernel uses to keep track of virtual
address ranges.

phys_addr does not necessarily need to be page-aligned. Ultimately,
the physical page it resides within will be mapped in full into a
page-aligned virtual address, but the returned address will be offset
to correspond exactly to the requested physical address.