Fix calculation of number of required pages
Previously, it was purely based on the size, but in practice because
the physical address does not need to be page-aligned, we may need one
additional page to cover the extra offset into the page. This change
calculates the range of physical addresses our mapping needs to cover,
then calculates the size based on that.