Fix section alignment when building program image
This should be a more robust way of building a particular program
segment in memory that avoids accidentally overrunning (as may have
happened before). Testing with a more verbose version verifies that
the program is loaded correctly (even though loading is a little
fragile at the moment) even if the sections in the ELF file are not
page-aligned.